refactoring: extract WorkdayUtils
