+++ lambda/src/main/java/com/lambda/IssueUpdater.java
@@ -1,16 +1,13 @@
 package com.lambda;
 
+import com.lambda.utils.WorkdayUtils;
 import java.text.DecimalFormat;
-import java.time.DayOfWeek;
 import java.time.Duration;
 import java.time.Instant;
 import java.time.LocalDateTime;
-import java.time.LocalTime;
 import java.time.ZoneId;
 import java.time.format.DateTimeParseException;
 import java.util.Optional;
-import java.util.Arrays;
-import java.util.List;
 
 import com.nulabinc.backlog4j.BacklogClient;
 import com.nulabinc.backlog4j.BacklogClientFactory;
@@ -27,9 +24,6 @@ public class IssueUpdater {
 
     private static final String CUSTOM_FIELD_STARTED_AT = "Started at";
     private static final ZoneId JST_ZONE = ZoneId.of("Asia/Tokyo");
-    private static final List<DayOfWeek> WEEKENDS = Arrays.asList(DayOfWeek.SATURDAY, DayOfWeek.SUNDAY);
-    private static final LocalTime WORK_START_TIME = LocalTime.of(9, 30); // Start at 09:30 JST
-    private static final LocalTime WORK_END_TIME = LocalTime.of(19, 30); // End at 17:30 ICT as 19:30 JST
 
 
     public IssueUpdater(final String apiKey) {
@@ -55,7 +49,7 @@ public class IssueUpdater {
                 // get elapsed time from the started time in hours
                 try {
                     LocalDateTime endAt = LocalDateTime.ofInstant(Instant.now(), JST_ZONE);
-                    elapsed = calculateWorkingHours(LocalDateTime.parse(startedAtValue), endAt);
+                    elapsed = new WorkdayUtils().calculateWorkingHours(LocalDateTime.parse(startedAtValue), endAt);
                 } catch (DateTimeParseException ex) {
                 }
             }
@@ -74,34 +68,6 @@ public class IssueUpdater {
             return null;
 
         return client.updateIssue(new UpdateIssueParams(issue.getId()).actualHours(actualHours));
-    }
-
-    private Duration calculateWorkingHours(final LocalDateTime start, final LocalDateTime end) {
-        LocalDateTime current = start;
-        Duration totalWorkingDuration = Duration.ZERO;
-
-        while (current.isBefore(end)) {
-            if (isWeekday(current)) {
-                LocalDateTime endOfDay = current.toLocalDate().atTime(WORK_END_TIME); 
-                
-                // Add hours for current day or until end time
-                LocalDateTime actualEnd = isSameDate(end, endOfDay) ? end : endOfDay;
-                totalWorkingDuration = totalWorkingDuration.plus(Duration.between(current, actualEnd));
-            }
-
-            // Move to the next day
-            current = current.plusDays(1).toLocalDate().atTime(WORK_START_TIME);
-        }
-
-        return totalWorkingDuration;
-    }
-
-    private boolean isWeekday(final LocalDateTime dateTime) {
-        return !WEEKENDS.contains(dateTime.getDayOfWeek());
-    }
-
-    private boolean isSameDate(final LocalDateTime dateTime1, final LocalDateTime dateTime2) {
-        return dateTime1.toLocalDate().equals(dateTime2.toLocalDate());
     }
 
     public Issue setStartedAt(final int issueId) {
